+++ frontend/lost/src/containers/DataSources/DSTable.jsx
@@ -12,7 +12,7 @@ import * as Notification from '../../components/Notification'
 import EditDSModal from './EditDSModal'
 // import { deleteFs } from '../../access/fb'
 import { CBadge } from '@coreui/react'
-import * as fbAPI from '../../actions/fb/fb_api'
+import * as fbAPI from '../../api/fb'
 import BaseContainer from '../../components/BaseContainer'
 import CoreDataTable from '../../components/CoreDataTable'
 import CoreIconButton from '../../components/CoreIconButton'
